Fixed uninstaller
Not backs up DLLs and exes that were already present during install.

+++ game_bridge_installer/GameBridgeInstaller/GUIHandler.cs
@@ -6,6 +6,7 @@ using System.IO;
 using System.Linq;
 using System.Security.Principal;
 using System.Text;
+using System.Text.RegularExpressions;
 using System.Threading.Tasks;
 
 namespace GameBridgeInstaller
@@ -15,11 +16,16 @@ namespace GameBridgeInstaller
         bool geo11FixFound = false;
         bool superDepth3DFixFound = false;
 
+        string gameExeName = "";
+        string gameExeNameWithoutExtension = "";
+        string gameExeFolderPath = "";
+
         string geo11FixPathPrefix = "Geo-11";
         string superDepth3DFixPathPrefix = "SuperDepth3D";
         string superDepth3DDefaultFixPathPrefix = "SuperDepth3D\\Default";
         string superDepth3DShaderPathPrefix = "SuperDepth3D\\ReShade-Shaders";
         string reshadeShaderPathPrefix = "reshade-shaders";
+        string srBackupExtensionPostfix = "SRBACKUP";
 
         string SRInstallpath = "";
         string SRAddonPath = "";
@@ -103,12 +109,15 @@ namespace GameBridgeInstaller
 
         public string InstallGameBridge(string pathToGameExe, GraphicsAPITypes graphicsApi)
         {
+            // Reset global variables
+            ResetGlobals();
+
             // Split exe name off from path.
-            string gameExeName = pathToGameExe.Substring(pathToGameExe.LastIndexOf("\\"), pathToGameExe.Length - pathToGameExe.LastIndexOf("\\"));
+            gameExeName = pathToGameExe.Substring(pathToGameExe.LastIndexOf("\\"), pathToGameExe.Length - pathToGameExe.LastIndexOf("\\"));
             // Cut the '.exe' part off the gameExeName.
-            string gameExeNameWithoutExtension = gameExeName.Split('.')[0];
+            gameExeNameWithoutExtension = gameExeName.Split('.')[0];
             // Cut the '.exe' part off the pathToGameExe. Pushed 1 index back to include the '\\' characters.
-            string gameExeFolderPath = pathToGameExe.Substring(0, pathToGameExe.LastIndexOf("\\") + 1);
+            gameExeFolderPath = pathToGameExe.Substring(0, pathToGameExe.LastIndexOf("\\") + 1);
 
             // First, copy all required SR files to the target path.
             try
@@ -199,13 +208,23 @@ namespace GameBridgeInstaller
                     // Copy all the files & Replaces any files with the same name
                     foreach (string newPath in Directory.GetFiles(geo11FixPathPrefix + gameExeNameWithoutExtension, "*.*", SearchOption.AllDirectories))
                     {
+                        // First see if the file already exists.
+                        // For now, we only want to back up .dll and .exe files to avoid confusion.
+                        if (File.Exists(newPath.Replace(geo11FixPathPrefix + gameExeNameWithoutExtension, gameExeFolderPath)) && newPath.Contains(".dll") || File.Exists(newPath.Replace(geo11FixPathPrefix + gameExeNameWithoutExtension, gameExeFolderPath)) && newPath.Contains(".exe"))
+                        {
+                            // File exists, rename it so we can restore it during uninstallation.
+                            File.Move(newPath.Replace(geo11FixPathPrefix + gameExeNameWithoutExtension, gameExeFolderPath), newPath.Replace(geo11FixPathPrefix + gameExeNameWithoutExtension, gameExeFolderPath) + srBackupExtensionPostfix, true);
+                        }
+
                         File.Copy(newPath, newPath.Replace(geo11FixPathPrefix + gameExeNameWithoutExtension, gameExeFolderPath), true);
                     }
                 }
                 catch (Exception ex)
                 {
                     // Todo: Implement cleaning up the fix if copying fails
-                    Console.WriteLine("WARNING: Cleaning up geo-11 fix files not yet implemented!");
+                    Console.WriteLine("WARNING: Cleaning up geo-11 fix files not finished yet!");
+                    Console.WriteLine("Restoring backed up Geo-11 files...");
+                    RestoreBackedUpFiles(gameExeFolderPath);
 
                     return "An exception occured while copying the geo-11 fix:\n\n" + ex.Message;
                 }
@@ -222,13 +241,23 @@ namespace GameBridgeInstaller
                     // Copy all the files & Replaces any files with the same name
                     foreach (string newPath in Directory.GetFiles(superDepth3DShaderPathPrefix, "*.*", SearchOption.AllDirectories))
                     {
+                        // First see if the file already exists.
+                        // For now, we only want to back up .dll and .exe files to avoid confusion.
+                        if (File.Exists(newPath.Replace(superDepth3DShaderPathPrefix, gameExeFolderPath + reshadeShaderPathPrefix)) && newPath.Contains(".dll") || File.Exists(newPath.Replace(superDepth3DShaderPathPrefix, gameExeFolderPath + reshadeShaderPathPrefix)) && newPath.Contains(".exe"))
+                        {
+                            // File exists, rename it so we can restore it during uninstallation.
+                            File.Move(newPath.Replace(superDepth3DShaderPathPrefix, gameExeFolderPath + reshadeShaderPathPrefix), newPath.Replace(superDepth3DShaderPathPrefix, gameExeFolderPath + reshadeShaderPathPrefix) + srBackupExtensionPostfix, true);
+                        }
+
                         File.Copy(newPath, newPath.Replace(superDepth3DShaderPathPrefix, gameExeFolderPath + reshadeShaderPathPrefix), true);
                     }
                 }
                 catch (Exception ex)
                 {
                     // Todo: Implement cleaning up the shader if copying fails
-                    Console.WriteLine("WARNING: Cleaning up SuperDepth3D shader files not yet implemented!");
+                    Console.WriteLine("WARNING: Cleaning up SuperDepth3D shader files not finished yet!");
+                    Console.WriteLine("Restoring backed up SuperDepth3D shader files...");
+                    RestoreBackedUpFiles(gameExeFolderPath);
 
                     return "An exception occured while copying the SuperDepth3D shader:\n\n" + ex.Message;
                 }
@@ -254,13 +283,23 @@ namespace GameBridgeInstaller
                     // Copy all the files & Replaces any files with the same name
                     foreach (string newPath in Directory.GetFiles(superDepth3DFixPathPrefix + gameExeNameWithoutExtension, "*.*", SearchOption.AllDirectories))
                     {
+                        // First see if the file already exists.
+                        // For now, we only want to back up .dll files to avoid confusion.
+                        if (File.Exists(newPath.Replace(superDepth3DFixPathPrefix + gameExeNameWithoutExtension, gameExeFolderPath)) && newPath.Contains(".dll"))
+                        {
+                            // File exists, rename it so we can restore it during uninstallation.
+                            File.Move(newPath.Replace(superDepth3DFixPathPrefix + gameExeNameWithoutExtension, gameExeFolderPath), newPath.Replace(superDepth3DFixPathPrefix + gameExeNameWithoutExtension, gameExeFolderPath) + srBackupExtensionPostfix, true);
+                        }
+
                         File.Copy(newPath, newPath.Replace(superDepth3DFixPathPrefix + gameExeNameWithoutExtension, gameExeFolderPath), true);
                     }
                 }
                 catch (Exception ex)
                 {
                     // Todo: Implement cleaning up the shader if copying fails
-                    Console.WriteLine("WARNING: Cleaning up SuperDepth3D fix files not yet implemented!");
+                    Console.WriteLine("WARNING: Cleaning up SuperDepth3D fix files not finished yet!");
+                    Console.WriteLine("Restoring backed up SuperDepth3D fix files...");
+                    RestoreBackedUpFiles(gameExeFolderPath);
 
                     return "An exception occured while copying the SuperDepth3D fix:\n\n" + ex.Message;
                 }
@@ -278,13 +317,23 @@ namespace GameBridgeInstaller
                     // Copy all the files & Replaces any files with the same name
                     foreach (string newPath in Directory.GetFiles(superDepth3DDefaultFixPathPrefix, "*.*", SearchOption.AllDirectories))
                     {
+                        // First see if the file already exists.
+                        // For now, we only want to back up .dll files to avoid confusion.
+                        if (File.Exists(newPath.Replace(superDepth3DDefaultFixPathPrefix, gameExeFolderPath)) && newPath.Contains(".dll"))
+                        {
+                            // File exists, rename it so we can restore it during uninstallation.
+                            File.Move(newPath.Replace(superDepth3DDefaultFixPathPrefix, gameExeFolderPath), newPath.Replace(superDepth3DDefaultFixPathPrefix, gameExeFolderPath) + srBackupExtensionPostfix, true);
+                        }
+
                         File.Copy(newPath, newPath.Replace(superDepth3DDefaultFixPathPrefix, gameExeFolderPath), true);
                     }
                 }
                 catch (Exception ex)
                 {
                     // Todo: Implement cleaning up the shader if copying fails
-                    Console.WriteLine("WARNING: Cleaning up SuperDepth3D default fix files not yet implemented!");
+                    Console.WriteLine("WARNING: Cleaning up SuperDepth3D default fix files not finished yet!");
+                    Console.WriteLine("Restoring backed up SuperDepth3D default fix files...");
+                    RestoreBackedUpFiles(gameExeFolderPath);
 
                     return "An exception occured while copying the default SuperDepth3D fix:\n\n" + ex.Message;
                 }
@@ -295,12 +344,15 @@ namespace GameBridgeInstaller
         public string UninstallGameBridge(string pathToGameExe, GraphicsAPITypes graphicsApi)
         {
             // Todo: Test this code
+            // Reset global variables
+            ResetGlobals();
+
             // Split exe name off from path.
-            string gameExeName = pathToGameExe.Substring(pathToGameExe.LastIndexOf("\\"), pathToGameExe.Length - pathToGameExe.LastIndexOf("\\"));
+            gameExeName = pathToGameExe.Substring(pathToGameExe.LastIndexOf("\\"), pathToGameExe.Length - pathToGameExe.LastIndexOf("\\"));
             // Cut the '.exe' part off the gameExeName.
-            string gameExeNameWithoutExtension = gameExeName.Split('.')[0];
+            gameExeNameWithoutExtension = gameExeName.Split('.')[0];
             // Cut the '.exe' part off the pathToGameExe. Pushed 1 index back to include the '\\' characters.
-            string gameExeFolderPath = pathToGameExe.Substring(0, pathToGameExe.LastIndexOf("\\") + 1);
+            gameExeFolderPath = pathToGameExe.Substring(0, pathToGameExe.LastIndexOf("\\") + 1);
 
             // Remove any files that were potentially copied.
             foreach (string dllName in SRDlls)
@@ -362,32 +414,77 @@ namespace GameBridgeInstaller
                 geo11FixFound = true;
             }
 
-            // Copy the geo-11 fix files (if Geo-11 fix is found)
+            // Remove the geo-11 fix files (if Geo-11 fix is found)
             if (geo11FixFound)
             {
-                
-                // Delete all of the subdirectories of the geo-11 fix
+                foreach (string filePath in Directory.GetFiles(geo11FixPathPrefix + gameExeNameWithoutExtension, "*.*", SearchOption.TopDirectoryOnly))
+                {
+                    // Delete from Geo-11 base path.
+                    DeleteFileByPathIfFoundInFix(filePath);
+                }
+
+                // Delete all of the files in all folders of the game exe path if they were previously installed as part of the Geo-11 fix.
+                foreach (string dirPath in Directory.GetDirectories(geo11FixPathPrefix + gameExeNameWithoutExtension, "*", SearchOption.AllDirectories))
+                {
+                    foreach (string filePath in Directory.GetFiles(dirPath, "*.*", SearchOption.TopDirectoryOnly))
+                    {
+                        DeleteFileByPathIfFoundInFix(filePath);
+                    }
+                }
+
+
+                // Todo: This is a lot of nested loops, performance can certainly be improved! Also you can make a method for this duplicated code!
+                // Delete all of the folders that are empty after deleting the files.
                 foreach (string dirPath in Directory.GetDirectories(geo11FixPathPrefix + gameExeNameWithoutExtension, "*", SearchOption.AllDirectories))
                 {
                     try
                     {
-                        Directory.Delete(dirPath.Replace(geo11FixPathPrefix + gameExeNameWithoutExtension, gameExeFolderPath), true);
+                        if (Directory.GetFiles(dirPath.Replace(geo11FixPathPrefix + gameExeNameWithoutExtension, gameExeFolderPath), "*.*", SearchOption.AllDirectories).Count() == 0)
+                        {
+                            Directory.Delete(dirPath.Replace(geo11FixPathPrefix + gameExeNameWithoutExtension, gameExeFolderPath), true);
+                        }
                     }
                     catch { }
                 }
-                // Delete all the files & Replaces any files with the same name
-                foreach (string filePath in Directory.GetFiles(geo11FixPathPrefix + gameExeNameWithoutExtension, "*.*", SearchOption.TopDirectoryOnly))
+            }
+
+            // Check for any SuperDepth3D fix files
+            if (Directory.Exists(superDepth3DFixPathPrefix + gameExeNameWithoutExtension))
+            {
+                superDepth3DFixFound = true;
+            }
+
+            // Remove the SupderDepth3D fix files (if SD3D fix is found)
+            if (superDepth3DFixFound)
+            {
+                // Check for any SuperDepth3D fix files, uninstall default reshadepreset.ini
+                foreach (string filePath in Directory.GetFiles(superDepth3DFixPathPrefix + gameExeNameWithoutExtension, "*.*", SearchOption.TopDirectoryOnly))
                 {
-                    // Check if file in game directory also exists inside Geo-11 fix directory.
-                    string selectedFileName = filePath.Substring(filePath.LastIndexOf("\\"), filePath.Length - filePath.LastIndexOf("\\")); 
-                    if (File.Exists(geo11FixPathPrefix + gameExeNameWithoutExtension + selectedFileName))
+                    // Delete from SD3D base path.
+                    DeleteFileByPathIfFoundInFix(filePath);
+                }
+
+                // Delete all of the files in all folders of the game exe path if they were previously installed as part of the SD3D fix.
+                foreach (string dirPath in Directory.GetDirectories(superDepth3DFixPathPrefix + gameExeNameWithoutExtension, "*", SearchOption.AllDirectories))
+                {
+                    foreach (string filePath in Directory.GetFiles(dirPath, "*.*", SearchOption.TopDirectoryOnly))
                     {
-                        try
-                        {
-                            File.Delete(filePath.Replace(geo11FixPathPrefix + gameExeNameWithoutExtension, gameExeFolderPath));
-                        }
-                        catch { }
+                        DeleteFileByPathIfFoundInFix(filePath);
                     }
+                }
+
+                // Todo: This is a lot of nested loops, performance can certainly be improved! Also you can make a method for this duplicated code
+                // Delete the directory if it's empty at this point.
+                foreach (string dirPath in Directory.GetDirectories(superDepth3DFixPathPrefix + gameExeNameWithoutExtension, "*", SearchOption.AllDirectories))
+                {
+                    try
+                    {
+                        if (Directory.GetFiles(dirPath.Replace(superDepth3DFixPathPrefix + gameExeNameWithoutExtension, gameExeFolderPath), "*.*", SearchOption.AllDirectories).Count() == 0)
+                        {
+                            Directory.Delete(dirPath.Replace(superDepth3DFixPathPrefix + gameExeNameWithoutExtension, gameExeFolderPath), true);
+                        }
+                    }
+                    catch { }
                 }
             }
 
@@ -401,6 +498,9 @@ namespace GameBridgeInstaller
                 }
                 catch { }
             }
+
+            // Restore backed up files
+            RestoreBackedUpFiles(gameExeFolderPath);
 
             return "";
         }
@@ -422,6 +522,63 @@ namespace GameBridgeInstaller
             {
                 return "";
             }
+        }
+
+        private void RestoreBackedUpFiles(string gameExeFolderPath)
+        {
+
+            DirectoryInfo hdDirectoryInWhichToSearch = new DirectoryInfo(gameExeFolderPath);
+            FileInfo[] filesInDir = hdDirectoryInWhichToSearch.GetFiles("*.*" + srBackupExtensionPostfix, SearchOption.AllDirectories);
+
+            foreach (FileInfo foundFile in filesInDir)
+            {
+                try
+                {
+                    string fullName = foundFile.FullName;
+                    if (fullName.Contains(srBackupExtensionPostfix))
+                    {
+                        foundFile.MoveTo(fullName.Replace(srBackupExtensionPostfix, "")); // Todo: Not sure if this places the files in the right place.
+                    }
+                } catch (Exception ex)
+                {
+                    Console.WriteLine("An exception occured while restoring the following file from backup:\n\n" + foundFile.FullName + "\n\nError message:\n\n" +  ex.Message);
+                }
+
+            }
+        }
+
+        private void DeleteFileByPathIfFoundInFix(string filePath)
+        {
+            // Check if file in game directory also exists inside Geo-11 fix directory.
+            // Make sure the file wasn't renamed with our backup postfix.
+            string selectedFileName = filePath.Substring(filePath.LastIndexOf("\\"), filePath.Length - filePath.LastIndexOf("\\"));
+            try
+            {
+                if (Directory.GetFiles(geo11FixPathPrefix + gameExeNameWithoutExtension, selectedFileName.Replace("\\", ""), SearchOption.AllDirectories).FirstOrDefault() != null && !selectedFileName.Contains(srBackupExtensionPostfix))
+                {
+                    File.Delete(filePath.Replace(geo11FixPathPrefix + gameExeNameWithoutExtension, gameExeFolderPath));
+                    return;
+                }
+            }
+            catch { }
+            try
+            {
+                if (Directory.GetFiles(superDepth3DFixPathPrefix + gameExeNameWithoutExtension, selectedFileName.Replace("\\", ""), SearchOption.AllDirectories).FirstOrDefault() != null  && !selectedFileName.Contains(srBackupExtensionPostfix))
+                {
+                    File.Delete(filePath.Replace(superDepth3DFixPathPrefix + gameExeNameWithoutExtension, gameExeFolderPath));
+                    return;
+                }
+            }
+            catch { }
+        }
+
+        private void ResetGlobals()
+        {
+            gameExeName = "";
+            gameExeNameWithoutExtension = "";
+            gameExeFolderPath = "";
+            geo11FixFound = false;
+            superDepth3DFixFound = false;
         }
     }
 }
